test: mock localStorage for all tests
